test: move standalone and distributed instances tests from "frontend" crate to "tests-integration"
