+++ src/frontend/src/instance/opentsdb.rs
@@ -35,95 +35,3 @@ impl OpentsdbProtocolHandler for Instance {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-
-    use common_query::Output;
-    use common_recordbatch::RecordBatches;
-    use itertools::Itertools;
-    use servers::query_handler::sql::SqlQueryHandler;
-    use session::context::QueryContext;
-
-    use super::*;
-    use crate::tests;
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_standalone_exec() {
-        let standalone = tests::create_standalone_instance("test_standalone_exec").await;
-        let instance = &standalone.instance;
-
-        test_exec(instance).await;
-    }
-
-    #[tokio::test(flavor = "multi_thread")]
-    async fn test_distributed_exec() {
-        let distributed = tests::create_distributed_instance("test_distributed_exec").await;
-        let instance = &distributed.frontend;
-
-        test_exec(instance).await;
-    }
-
-    async fn test_exec(instance: &Arc<Instance>) {
-        let ctx = QueryContext::arc();
-        let data_point1 = DataPoint::new(
-            "my_metric_1".to_string(),
-            1000,
-            1.0,
-            vec![
-                ("tagk1".to_string(), "tagv1".to_string()),
-                ("tagk2".to_string(), "tagv2".to_string()),
-            ],
-        );
-        // should create new table "my_metric_1" directly
-        let result = instance.exec(&data_point1, ctx.clone()).await;
-        assert!(result.is_ok());
-
-        let data_point2 = DataPoint::new(
-            "my_metric_1".to_string(),
-            2000,
-            2.0,
-            vec![
-                ("tagk2".to_string(), "tagv2".to_string()),
-                ("tagk3".to_string(), "tagv3".to_string()),
-            ],
-        );
-        // should create new column "tagk3" directly
-        let result = instance.exec(&data_point2, ctx.clone()).await;
-        assert!(result.is_ok());
-
-        let data_point3 = DataPoint::new("my_metric_1".to_string(), 3000, 3.0, vec![]);
-        // should handle null tags properly
-        let result = instance.exec(&data_point3, ctx.clone()).await;
-        assert!(result.is_ok());
-
-        let output = instance
-            .do_query(
-                "select * from my_metric_1 order by greptime_timestamp",
-                Arc::new(QueryContext::new()),
-            )
-            .await
-            .remove(0)
-            .unwrap();
-        match output {
-            Output::Stream(stream) => {
-                let recordbatches = RecordBatches::try_collect(stream).await.unwrap();
-                let pretty_print = recordbatches.pretty_print().unwrap();
-                let expected = vec![
-                    "+---------------------+----------------+-------+-------+-------+",
-                    "| greptime_timestamp  | greptime_value | tagk1 | tagk2 | tagk3 |",
-                    "+---------------------+----------------+-------+-------+-------+",
-                    "| 1970-01-01T00:00:01 | 1.0            | tagv1 | tagv2 |       |",
-                    "| 1970-01-01T00:00:02 | 2.0            |       | tagv2 | tagv3 |",
-                    "| 1970-01-01T00:00:03 | 3.0            |       |       |       |",
-                    "+---------------------+----------------+-------+-------+-------+",
-                ]
-                .into_iter()
-                .join("\n");
-                assert_eq!(pretty_print, expected);
-            }
-            _ => unreachable!(),
-        };
-    }
-}
