+++ src/frontend/src/lib.rs
@@ -18,7 +18,7 @@
 pub mod catalog;
 pub mod datanode;
 pub mod error;
-mod expr_factory;
+pub mod expr_factory;
 pub mod frontend;
 pub mod grpc;
 pub mod influxdb;
@@ -32,11 +32,4 @@ pub mod prometheus;
 mod script;
 mod server;
 pub mod statement;
-mod table;
-#[cfg(test)]
-mod tests;
-
-#[cfg(test)]
-// allowed because https://docs.rs/rstest_reuse/0.5.0/rstest_reuse/#use-rstest_reuse-at-the-top-of-your-crate
-#[allow(clippy::single_component_path_imports)]
-use rstest_reuse;
+pub mod table;
